requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.27.2
python-multipart==0.0.17
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Redis Cache
redis==5.2.1
aiocache==0.12.3

# Data Processing
pandas==2.2.3
numpy==2.2.0

# Scheduling
apscheduler==3.10.4
croniter==3.0.4

# Analytics & Visualization
matplotlib==3.9.3
seaborn==0.13.2
plotly==5.24.1

# Notifications
aiobotocore==2.16.1  # For SES
twilio==9.3.8  # For SMS
firebase-admin==6.6.0  # For push notifications

# Async Support
aiofiles==24.1.0
asyncio==3.4.3

# Authentication
python-jose[cryptography]==3.3.0

# Logging & Monitoring
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0

# Utilities
orjson==3.10.12
tenacity==9.0.0
python-dateutil==2.9.0.post0

# Development
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx-mock==0.4.0
black==24.10.0
ruff==0.8.3
mypy==1.13.0
pre-commit==4.0.1